+++ src/app/(common)/search/search.module.scss
