+++ src/app/(sign)/signIn/findPassword/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import FindPassword from '../../_components/FindPassword'
+
+const page = () => {
+  return <FindPassword />
+}
+
+export default page
