+++ src/app/(sign)/signIn/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import SignIn from '../_components/SignIn'
+
+const page = () => {
+  return <SignIn />
+}
+
+export default page
