+++ src/app/(sign)/signUp/page.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import SignUp from '../_components/SignUp'
+
+const page = () => {
+  return <SignUp />
+}
+
+export default page
