+++ src/app/detail/[id]/RelatedVedio.tsx
@@ -4,7 +4,7 @@ import React, { useState, useEffect, useCallback } from 'react'
 import Link from 'next/link'
 import { IYoutubeItem } from '@/type/YoutubeApiResponse'
 import formatRelativeDate from '@/utils/relativeDate'
-import ScrollBtn from '@/components/ScrollBtn'
+import ScrollBtn from '@/app/(common)/_components/ScrollBtn'
 import useYoutubeDataRequest from '@/hooks/useYoutubeApiRequest'
 import Image from 'next/image'
 import styles from './detail.module.scss'
