+++ src/app/page.module.scss
@@ -53,7 +53,7 @@
     width: 60%;
     transition: all 0.3s ease-in-out;
     padding: 10.25rem 16.25rem 4.25rem 12.25rem;
-    
+
     h2 {
       color: #da3017;
     }
@@ -144,7 +144,7 @@
       left: 15%;
       width: 70%;
       height: 70%;
-      background-color: #FE5D65;
+      background-color: #fe5d65;
     }
     .counterContent {
       position: relative;
@@ -179,9 +179,10 @@
 
   .videoCard {
     font-size: 0;
+    padding: 0.2rem;
     position: relative;
     margin: 10px;
-    border-radius: 20px;
+    border-radius: 5px;
     overflow: hidden;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
     transition: all 0.3s ease-in-out;
@@ -207,9 +208,9 @@
       right: 5%;
       font-size: 1.25rem;
       font-weight: bold;
-      color: #DA3017;
+      color: #da3017;
       border-radius: 1.5rem;
-      background-color: #FFDCDC;
+      background-color: #ffdcdc;
       padding: 0.3125rem 0.4315rem;
     }
   }
@@ -246,7 +247,7 @@
   }
   .date {
     font-size: 1rem;
-    color: #DA3017;
+    color: #da3017;
   }
 }
 
@@ -302,7 +303,7 @@
         mix-blend-mode: luminosity;
       }
     }
-    
+
     .candyStick {
       width: 5.375rem;
       height: auto;
@@ -311,7 +312,8 @@
       transform: rotate(-26deg);
     }
   }
-  h2, p {
+  h2,
+  p {
     text-align: center;
   }
 
@@ -369,7 +371,7 @@
         height: 6.25rem;
         z-index: 1;
         text-align: center;
-        background: no-repeat center/cover url("/assets/ball.png");
+        background: no-repeat center/cover url('/assets/ball.png');
       }
       div {
         width: 1.875rem;
@@ -388,10 +390,18 @@
       display: flex;
       justify-content: space-between;
       align-items: center;
-      background: #fff linear-gradient(180deg, rgba(0, 0, 0, 0.20) 0%, rgba(247, 247, 247, 0.20) 13.02%, rgba(255, 255, 255, 0.20) 91.15%, rgba(0, 0, 0, 0.20) 100%);
-      border-top: 9px solid #1C4A31;
-      border-bottom: 9px solid #1C4A31;
-      &::before, &::after {
+      background: #fff
+        linear-gradient(
+          180deg,
+          rgba(0, 0, 0, 0.2) 0%,
+          rgba(247, 247, 247, 0.2) 13.02%,
+          rgba(255, 255, 255, 0.2) 91.15%,
+          rgba(0, 0, 0, 0.2) 100%
+        );
+      border-top: 9px solid #1c4a31;
+      border-bottom: 9px solid #1c4a31;
+      &::before,
+      &::after {
         content: '';
         display: block;
         position: absolute;
@@ -405,7 +415,6 @@
         left: 0;
         // transform: rotate(45deg) translateX(-50%);
 
-        
         border-bottom: 2.09375rem solid transparent;
         border-top: 2.09375rem solid transparent;
         border-left: 2.09375rem solid #146238;
@@ -434,8 +443,9 @@
 
 .survey {
   background-color: var(--color-main-ground-two-bg);
-  h2, p {
-    color: #0C412F;
+  h2,
+  p {
+    color: #0c412f;
   }
   .toggleList {
     margin-top: 4.375rem;
@@ -520,7 +530,6 @@
     }
   }
 }
-
 
 .Loading {
   text-align: center;
